backend/requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
Django==4.0.3
django-appconf==1.0.5
django-cors-headers==3.11.0
django-extensions==3.1.5
django-imagekit==4.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
numpy==1.22.3
pandas==1.4.1
pilkit==2.0
Pillow==9.0.1
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.1
